ternyata format tipe data antar table relasi tidak sama, harusnya big integer eh ternyata integer

+++ database/migrations/2022_07_12_020152_create_request_produksi_table.php
@@ -15,7 +15,7 @@ class CreateRequestProduksiTable extends Migration
     {
         Schema::create('request_produksi', function (Blueprint $table) {
             $table->increments('id_reqproduksi');
-            $table->unsignedInteger('id');
+            $table->unsignedBigInteger('id');
             $table->foreign('id')
                     ->references('id')
                     ->on('users')
